envファイルでURLとポート指定、websocketをbackgroundから接続する
