+++ .env.default
@@ -1,0 +1,2 @@
+SERVER_URL=http://localhost:80
+PORT=80
