+++ docker-compose.yml
@@ -1,10 +1,12 @@
-version: '3'
+version: '2'
 services:
   web:
     image: chimerast/niconico-speenya:develop
+    env_file:
+      - .env
     environment:
-      SERVER_URL: http://localhost:8080
+      SERVER_URL: http://52.69.180.141:80
     ports:
-      - 8080:8080
+      - 80:8080
     volumes:
       - ./data/:/app/data/
