+++ webpack.config.ts
@@ -27,7 +27,7 @@ const config: Configuration = {
   },
   plugins: [
     new DefinePlugin({
-      'process.env.SERVER_URL': JSON.stringify(process.env.SERVER_URL || 'http://localhost:3000'),
+      'process.env.SERVER_URL': JSON.stringify(process.env.SERVER_URL || 'http://localhost:80'),
     }),
     new CopyPlugin({
       patterns: [
